moving the Serach functionality to the Header component
