--- src/features/Search/styled.js
@@ -1,5 +1,0 @@
-import styled from "styled-components";
-
-export const SearchContainer = styled.form`
-  margin-top: 40px;
-`;
