+++ src/rootSaga.js
@@ -5,7 +5,7 @@ import { watchFetchPerson } from "./features/people/PersonPage/personSaga";
 import { watchFetchPopularPeople } from "./features/people/popularPeopleSaga";
 import { watchFetchMovieDetails } from "./features/movies/MoviePage/movieDetailsSaga";
 import { watchFetchMovie } from "./features/movies/MoviePage/movieSaga";
-import { watchFetchSearchQuery } from "./features/Search/searchSaga";
+import { watchFetchSearchQuery } from "./common/Header/searchSaga";
 import { watchFetchGenres } from "./common/tiles/MovieTile/Genre/genreSaga";
 
 export default function* rootSaga() {
